Split up modelTaskQueue.ts into three files
